[Add] - Delete folder when user deleted

+++ controllers/controller.user.js
@@ -28,6 +28,7 @@ const updateUser = catchAsync(async (req, res) => {
 	const { userId } = req.params;
 	const data = { ...req.body };
 	const user = await User.getUserById(userId);
+	if (!user) throw new ApiError(httpStatus.NOT_FOUND, "User not found");
 	if (req.file) {
 		if (user.photo.url && user.photo.path) File.deleteFile(user.photo.path);
 		data.photo = File.uploadFile(req.file, { userId, folder: "photo" });
@@ -37,7 +38,9 @@ const updateUser = catchAsync(async (req, res) => {
 });
 
 const deleteUser = catchAsync(async (req, res) => {
-	await User.deleteUserById(req.params.userId);
+	const { userId } = req.params;
+	await User.deleteUserById(userId);
+	File.deleteFolder(userId);
 	res.status(httpStatus.NO_CONTENT).json();
 });
